newer clang complains about abs of unsigned

+++ schism/page_instruments.c
@@ -1565,8 +1565,8 @@ static int _env_handle_mouse(struct key_event *k, song_envelope_t *env, int *cur
                         if (y > 21) y--;
                         y = 206 - y;
 
-                        dx = abs(x - k->fx);
-                        dy = abs(y - k->fy);
+                        dx = abs(x - (int) k->fx);
+                        dy = abs(y - (int) k->fy);
                         dist = i_sqrt((dx*dx)+(dy*dy));
                         if (best_dist_node == -1 || dist < best_dist) {
                                 if (dist <= 5) {
